Refactoring: KeyValueStorageBase
Put code from BrowserStorage and NodeFileStorage into common base class
